Update slides based on instruction additions for Python 3 (#39)
* Add indentation for for loop

For loop must be indented under with statement.

* Use 'next' instead of '.next'

The standard way to get the next item out of an iterator is to
use the 'next' funciton. This is especially important going between
python 2 and 3, since in python 3 iterators do not have '.next'
and have '.__next__' instead.

* minor suggested edits to slides

One erratum:
* float() instead of floats()
Plus a few minor items:
*Updated years
*Minor syntax suggestions

* carry updates over to French slides

* Add 'raw_input' alternative for python 3

Although the class requirements say python 2, a lot of folks
end up using python 3 instead, so added the alternative needed
instead of `raw_input` (there doesn't seem to be an alternative
that works for both).